# stage2: optimize AstGen.callExpr