<details>
<summary>tldr</summary>

| type | average improvement |
| --- | --- |
| debug no cache | ~549ms |
| debug cached | ~1.7ms |
| release no cache |  ~196ms |
| release cached | ~0.1ms |

</details>

While working on #10068 I happened to notice a sizeable number of short term allocations in this function.

I collected data regarding the number of arguments in call expressions by compiling a file with only an empty main function with no local or global cache and stashing the data from `AstGen.callExpr` which resulted in:

| Length | Count | Percentage |
| --- | --- | --- |
| 1 | 19480 | 45.80% |
| 2 | 7962 | 18.72% |
| 0 | 7140 | 16.79% |
| 3 | 4152 | 9.76% |
| 4 | 2042 | 4.80% |
| 5 | 1036 | 2.44% |
| 6 | 332 | 0.78% |
| 7 | 213 | 0.50% |
| 10 | 158 | 0.37% |
| 11 | 5 | 0.01% |
| 9 | 5 | <0.01% |
| 8 | 3 | <0.01% |
| 12 | 2 | <0.01% |

Using this information to prevent allocations for call expressions with less than 5 arguments seems to show good improvement (at least for compiling a file with nothing but an empty main)

the tldr at the top contains a summary of the below information

> :warning: **the below no cache timings include the time to delete the cached files and folders**

## debug no cache
| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `rm -rf zig-cache && rm -rf simple && rm -rf ~/.cache/zig && ~/src/zig/zig-out/changed/bin/zig build-exe simple.zig` | 2.364 ± 0.020 | 2.326 | 2.394 | 1.00 |
| `rm -rf zig-cache && rm -rf simple && rm -rf ~/.cache/zig && ~/src/zig/zig-out/normal/bin/zig build-exe simple.zig` | 2.913 ± 0.029 | 2.877 | 2.953 | 1.23 ± 0.02 |

## debug cached
| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `~/src/zig/zig-out/changed/bin/zig build-exe simple.zig` | 66.4 ± 4.7 | 55.1 | 76.7 | 1.00 |
| `~/src/zig/zig-out/normal/bin/zig build-exe simple.zig` | 68.1 ± 5.9 | 55.2 | 81.0 | 1.02 ± 0.11 |

## release no cache
| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `rm -rf zig-cache && rm -rf simple && rm -rf ~/.cache/zig && ~/src/zig/zig-out/changed/bin/zig build-exe simple.zig` | 951.9 ± 12.8 | 934.3 | 973.1 | 1.00 |
| `rm -rf zig-cache && rm -rf simple && rm -rf ~/.cache/zig && ~/src/zig/zig-out/normal/bin/zig build-exe simple.zig` | 1148.0 ± 15.1 | 1133.3 | 1179.4 | 1.21 ± 0.02 |

## release cached
| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `~/src/zig/zig-out/changed/bin/zig build-exe simple.zig` | 18.2 ± 2.2 | 12.2 | 24.4 | 1.01 |
| `~/src/zig/zig-out/normal/bin/zig build-exe simple.zig` | 18.3 ± 2.7 | 13.1 | 24.7 | 1.01 ± 0.19 |