@andrewrk 2021-10-31
BTW I'm also working on rebooting https://github.com/ziglang/gotta-go-fast/ so that we can track improvements like this in a visual way over time :-)

Hopefully within a couple weeks, with some backfilled data, we will be able to look at this commit in a pretty graph and see a bump up in perf!